pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "postgis-property-search-demo-py"
version = "1.0.0"
description = "Educational repository for Real Estate/Property Search with AWS infrastructure"
authors = [
    {name = "Learning Team", email = "learning@example.com"},
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Education",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Topic :: Database",
    "Topic :: Scientific/Engineering :: GIS",
]

dependencies = [
    "psycopg2-binary>=2.9.7",
    "geopandas>=0.14.0",
    "shapely>=2.0.1",
    "pandas>=2.1.0",
    "numpy>=1.24.3",
    "sqlalchemy>=2.0.20",
    "alembic>=1.12.0",
    "flask>=2.3.3",
    "flask-cors>=4.0.0",
    "gunicorn>=21.2.0",
    "boto3>=1.28.62",
    "structlog>=23.1.0",
    "prometheus-client>=0.17.1",
    "redis>=4.6.0",
    "celery>=5.3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.2",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "flake8>=6.0.0",
    "mypy>=1.5.1",
    "pre-commit>=3.3.3",
]

[project.urls]
Homepage = "https://github.com/example/postgis-property-search-demo-py"
Documentation = "https://github.com/example/postgis-property-search-demo-py#readme"
Repository = "https://github.com/example/postgis-property-search-demo-py"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers --cov=src --cov-report=term-missing"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
